// src/lib/api/generated/models/createPackDtoPriceMode.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Beauty Pack Recommendation API
 * Backend API for beauty-pack personalization, recommendations, COD orders, and administration
 * OpenAPI spec version: 1.0.0
 */

export type CreatePackDtoPriceMode = typeof CreatePackDtoPriceMode[keyof typeof CreatePackDtoPriceMode];


export const CreatePackDtoPriceMode = {
  FIXED: 'FIXED',
  SUM_ITEMS: 'SUM_ITEMS',
  SUM_ITEMS_WITH_DISCOUNT: 'SUM_ITEMS_WITH_DISCOUNT',
} as const;