// src/lib/api/generated/models/createPackDtoStatus.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Beauty Pack Recommendation API
 * Backend API for beauty-pack personalization, recommendations, COD orders, and administration
 * OpenAPI spec version: 1.0.0
 */

export type CreatePackDtoStatus = typeof CreatePackDtoStatus[keyof typeof CreatePackDtoStatus];


export const CreatePackDtoStatus = {
  DRAFT: 'DRAFT',
  ACTIVE: 'ACTIVE',
  ARCHIVED: 'ARCHIVED',
} as const;